Refactor code into a Python module
Move classes into their own files in a module
Adopt standard catkin support for the module so that other packages can use the wrappers directly
