+++ src/respeaker_ros/__init__.py
@@ -1,0 +1,3 @@
+from .audio import RespeakerAudio
+from .interface import RespeakerInterface
+from .node import RespeakerNode
